requirements.txt
cryptography==42.0.2
Pillow==9.0.1
numpy==1.21.6